Add support for generating auctionmark databases with a larger scalefactor

+++ src/com/oltpbenchmark/benchmarks/auctionmark/AuctionMarkLoader.java
@@ -604,7 +604,7 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
         
         private final LinkedBlockingDeque<T> queue = new LinkedBlockingDeque<T>();
         private T current;
-        private short currentCounter;
+        private int currentCounter;
         private boolean stop = false;
         private final String sourceTableName;
 
@@ -613,8 +613,8 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
             this.sourceTableName = sourceTableName;
         }
         
-        protected abstract short getElementCounter(T t);
-        protected abstract int populateRow(T t, Object[] row, short remaining);
+        protected abstract int getElementCounter(T t);
+        protected abstract int populateRow(T t, Object[] row, int remaining);
         
         public void stopWhenEmpty() {
             if (LOG.isDebugEnabled())
@@ -949,11 +949,11 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
                                                     AuctionMarkConstants.USER_MAX_ATTRIBUTES, 1.001);
         }
         @Override
-        protected short getElementCounter(UserId user_id) {
-            return (short)(randomNumUserAttributes.nextInt());
+        protected int getElementCounter(UserId user_id) {
+            return (int)(randomNumUserAttributes.nextInt());
         }
         @Override
-        protected int populateRow(UserId user_id, Object[] row, short remaining) {
+        protected int populateRow(UserId user_id, Object[] row, int remaining) {
             int col = 0;
             
             // UA_ID
@@ -991,8 +991,8 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
         }
         
         @Override
-        protected short getElementCounter(UserId user_id) {
-            return (short)(user_id.getItemCount());
+        protected int getElementCounter(UserId user_id) {
+            return (int)(user_id.getItemCount());
         }
 
         @Override
@@ -1004,7 +1004,7 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
             } // FOR
         }
         @Override
-        protected int populateRow(UserId seller_id, Object[] row, short remaining) {
+        protected int populateRow(UserId seller_id, Object[] row, int remaining) {
             int col = 0;
             
             ItemId itemId = new ItemId(seller_id, remaining);
@@ -1030,8 +1030,8 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
             assert(p != null);
 
             // Calculate the number of bids and watches for this item
-            short numBids = (short)p.first.nextInt();
-            short numWatches = (short)p.second.nextInt();
+            int numBids = p.first.nextInt();
+            int numWatches = p.second.nextInt();
             
             // Create the ItemInfo object that we will use to cache the local data 
             // for this item. This will get garbage collected once all the derivative
@@ -1044,7 +1044,7 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
             itemInfo.numImages = (short) profile.randomNumImages.nextInt();
             itemInfo.numAttributes = (short) profile.randomNumAttributes.nextInt();
             itemInfo.numBids = numBids;
-            itemInfo.numWatches = numWatches;
+            itemInfo.numWatches = (short) numWatches;
             
             // The auction for this item has already closed
             if (itemInfo.endDate.getTime() <= profile.getLoaderStartTime().getTime()) {
@@ -1139,11 +1139,12 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
                   AuctionMarkConstants.TABLENAME_ITEM);
         }
         @Override
-        public short getElementCounter(LoaderItemInfo itemInfo) {
+        public int getElementCounter(LoaderItemInfo itemInfo) {
             return itemInfo.numImages;
         }
         @Override
-        protected int populateRow(LoaderItemInfo itemInfo, Object[] row, short remaining) {
+        protected int populateRow(LoaderItemInfo itemInfo, Object[] row,
+                int remaining) {
             int col = 0;
 
             // II_ID
@@ -1169,11 +1170,12 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
                   AuctionMarkConstants.TABLENAME_GLOBAL_ATTRIBUTE_VALUE);
         }
         @Override
-        public short getElementCounter(LoaderItemInfo itemInfo) {
+        public int getElementCounter(LoaderItemInfo itemInfo) {
             return itemInfo.numAttributes;
         }
         @Override
-        protected int populateRow(LoaderItemInfo itemInfo, Object[] row, short remaining) {
+        protected int populateRow(LoaderItemInfo itemInfo, Object[] row,
+                int remaining) {
             int col = 0;
             GlobalAttributeValueId gav_id = profile.getRandomGlobalAttributeValue();
             assert(gav_id != null);
@@ -1203,11 +1205,12 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
                   AuctionMarkConstants.TABLENAME_ITEM);
         }
         @Override
-        public short getElementCounter(LoaderItemInfo itemInfo) {
+        public int getElementCounter(LoaderItemInfo itemInfo) {
             return (itemInfo.purchaseDate != null ? itemInfo.numComments : 0);
         }
         @Override
-        protected int populateRow(LoaderItemInfo itemInfo, Object[] row, short remaining) {
+        protected int populateRow(LoaderItemInfo itemInfo, Object[] row,
+                int remaining) {
             int col = 0;
 
             // IC_ID
@@ -1254,11 +1257,12 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
                   AuctionMarkConstants.TABLENAME_ITEM);
         }
         @Override
-        public short getElementCounter(LoaderItemInfo itemInfo) {
-            return ((short)itemInfo.numBids);
+        public int getElementCounter(LoaderItemInfo itemInfo) {
+            return ((int)itemInfo.numBids);
         }
         @Override
-        protected int populateRow(LoaderItemInfo itemInfo, Object[] row, short remaining) {
+        protected int populateRow(LoaderItemInfo itemInfo, Object[] row,
+                int remaining) {
             int col = 0;
             assert(itemInfo.numBids > 0);
             
@@ -1356,11 +1360,12 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
                   AuctionMarkConstants.TABLENAME_ITEM_BID);
         }
         @Override
-        public short getElementCounter(LoaderItemInfo itemInfo) {
-            return (short)(itemInfo.getBidCount() > 0 ? 1 : 0);
+        public int getElementCounter(LoaderItemInfo itemInfo) {
+            return (int)(itemInfo.getBidCount() > 0 ? 1 : 0);
         }
         @Override
-        protected int populateRow(LoaderItemInfo itemInfo, Object[] row, short remaining) {
+        protected int populateRow(LoaderItemInfo itemInfo, Object[] row,
+                int remaining) {
             int col = 0;
             LoaderItemInfo.Bid bid = itemInfo.getLastBid();
             assert(bid != null) : "No bids?\n" + itemInfo;
@@ -1394,11 +1399,12 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
                   AuctionMarkConstants.TABLENAME_ITEM_BID);
         }
         @Override
-        public short getElementCounter(LoaderItemInfo itemInfo) {
-            return (short)(itemInfo.getBidCount() > 0 && itemInfo.purchaseDate != null ? 1 : 0);
+        public int getElementCounter(LoaderItemInfo itemInfo) {
+            return (int)(itemInfo.getBidCount() > 0 && itemInfo.purchaseDate != null ? 1 : 0);
         }
         @Override
-        protected int populateRow(LoaderItemInfo itemInfo, Object[] row, short remaining) {
+        protected int populateRow(LoaderItemInfo itemInfo, Object[] row,
+                int remaining) {
             int col = 0;
             LoaderItemInfo.Bid bid = itemInfo.getLastBid();
             assert(bid != null) : itemInfo;
@@ -1437,11 +1443,12 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
         }
 
         @Override
-        protected short getElementCounter(LoaderItemInfo.Bid bid) {
-            return (short)((bid.buyer_feedback ? 1 : 0) + (bid.seller_feedback ? 1 : 0));
+        protected int getElementCounter(LoaderItemInfo.Bid bid) {
+            return (int)((bid.buyer_feedback ? 1 : 0) + (bid.seller_feedback ? 1 : 0));
         }
         @Override
-        protected int populateRow(LoaderItemInfo.Bid bid, Object[] row, short remaining) {
+        protected int populateRow(LoaderItemInfo.Bid bid, Object[] row,
+                int remaining) {
             int col = 0;
 
             boolean is_buyer = false;
@@ -1480,11 +1487,12 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
                   AuctionMarkConstants.TABLENAME_ITEM_BID);
         }
         @Override
-        public short getElementCounter(LoaderItemInfo itemInfo) {
-            return (short)(itemInfo.getBidCount() > 0 && itemInfo.purchaseDate != null ? 1 : 0);
+        public int getElementCounter(LoaderItemInfo itemInfo) {
+            return (int)(itemInfo.getBidCount() > 0 && itemInfo.purchaseDate != null ? 1 : 0);
         }
         @Override
-        protected int populateRow(LoaderItemInfo itemInfo, Object[] row, short remaining) {
+        protected int populateRow(LoaderItemInfo itemInfo, Object[] row,
+                int remaining) {
             int col = 0;
             LoaderItemInfo.Bid bid = itemInfo.getLastBid();
             assert(bid != null) : itemInfo;
@@ -1522,11 +1530,12 @@ public class AuctionMarkLoader extends Loader<AuctionMarkBenchmark> {
                   AuctionMarkConstants.TABLENAME_ITEM_BID);
         }
         @Override
-        public short getElementCounter(LoaderItemInfo itemInfo) {
+        public int getElementCounter(LoaderItemInfo itemInfo) {
             return (itemInfo.numWatches);
         }
         @Override
-        protected int populateRow(LoaderItemInfo itemInfo, Object[] row, short remaining) {
+        protected int populateRow(LoaderItemInfo itemInfo, Object[] row,
+                int remaining) {
             int col = 0;
             
             // Make it more likely that a user that has bid on an item is watching it
